there should be boolean values only on assertions
