+++ apex-checks/src/test/java/org/fundacionjala/enforce/sonarqube/apex/checks/testrelated/BooleanParamentersOnAssertTest.java
@@ -1,0 +1,61 @@
+/*
+ * The MIT License
+ *
+ * Copyright 2016 Fundacion Jala.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package org.fundacionjala.enforce.sonarqube.apex.checks.testrelated;
+
+import com.sonar.sslr.api.AstNode;
+import java.io.File;
+import java.util.LinkedList;
+import java.util.List;
+import org.fundacionjala.enforce.sonarqube.apex.ApexAstScanner;
+import org.junit.Test;
+import org.sonar.squidbridge.api.SourceFile;
+import org.sonar.squidbridge.checks.CheckMessagesVerifier;
+
+import static org.mockito.Mockito.mock;
+
+public class BooleanParamentersOnAssertTest {
+
+    private BooleanParamentersOnAssertCheck check;
+
+    @Test
+    public void testBooleanParametersOnAssert() {
+        File file = new File("src/test/resources/checks/testsClassAssertions.cls");
+        check = new BooleanParamentersOnAssertCheck();
+        SourceFile sourceFile = ApexAstScanner.scanFile(file, check);
+        CheckMessagesVerifier.verify(sourceFile.getCheckMessages())
+                .next().atLine(82).withMessage("Method testMethodWithNoBooleanValueOnAssert"
+                + "has a wrong System.assert parameter, it should be a boolean one")
+                .next().atLine(90).withMessage("Method testMethodWithExpressionValueOnAssert"
+                + "has a wrong System.assert parameter, it should be a boolean one")
+                .noMore();
+    }
+
+    @Test
+    public void testHasBooleanParameters() {
+        AstNode primarySuffix = mock(AstNode.class);
+        List<AstNode> primarySuffixes = new LinkedList<>();
+        primarySuffixes.add(primarySuffix);
+    }
+
+}
